requirements.txt
streamlit
google-generativeai
python-dotenv
gTTS
moviepy
fpdf2
requests
Pillow
